geminipro-deltalake/requirements.txt
langchain-core
google-generativeai
langchain
streamlit
python-dotenv